Fix zone deletion.
When deleting a zone, the old version only removed it from the DOM, but not from
the underlying data structure that is sent to the server when the save button is
pressed.  This resulted in zones not actually being deleted.  This commit fixes
this issue.  Also fix a typo in a variable name while we are here.