Use numbers instead of letters to identify sample zones
Prior to this patch, the sample zones in the default dataset would use
letters to identify zones ("Zone A"), but zones added to the default set
would use numbers ("Zone 3"). This changes the sample zones to use
numbers too.

For consistency and maintenability, the change was also applied to all
tests, even when the test would still have worked with the letter
format.