Ensure old problems work on new version.
Recent versions of the drag and drop v2 xblock require the 'correct'
field to be present in items stored in user state, but the 'correct'
field did not exist before assessment mode has been implemented and so
problems created and completed before the introduction of assessment
mode lack the 'correct' field.

This patch adds the 'correct' field to item state at load time.
In standard mode, only items that are placed in correct zone are stored
in item state, so we can assume that any item stored in the state that
does not have the 'correct' field is 'correct'.