When dragging, scroll target image as needed
If the target image is larger than the screen, scroll the target area as
necessary while dragging to allow the user to reach the drop zone.

When setting up the threshold areas which trigger scrolling, fixed
headers that overlay the top of the image need to be considered, so I've
added a check for those.

Tested on desktop browsers, mobile Safari, Chrome on Android, and iOS
native apps.