(BB-1079) Fix bug affecting dragging on mobile (#204)
When having a large draggable item, the limits are not calculated correctly and the XBlock doesn't let the user to drag the item all the way to the right. This happens because the limit is calculated once when the drag is initiated. But when the item is moved the actual div that gets moved is a different one created when an item is marked as dragged and a different style is applied to it making the width only 30% of the original. This makes the limit wrong. The solution is to calculate the limits by using the actual div that's being dragged.

This commit calculates the limits correctly and it only does it once by using a sort of memoization for performance. We also bump the version to 2.2.3.