fix: Properly limit 'feedback.opened' events
Only send ‘feedback.opened’ if there are elements in feedback array.

See https://openedx.atlassian.net/browse/TNL-7184 for complete details and more background.

Long story short, if a drag and drop item doesn't have a success or error feedback message, the backend sends an empty array of feedback, which the frontend wasn't consuming properly.  This PR allows the frontend to interpret an empty array as not worthy of sending a feedback.opened event.