feat: use platform translations as primary and local as secondary
This changes the way the JS file access translation for strings using
the Django's gettext & ngettext functions. Instead of limiting fully to
the local translations, this uses the platform's translations as the
primary source and falls back to the local translations if the string is
not available.

This allows for overriding of the DnD's strings from other places like
the themes by patching the Django's global catalog.